[Proposal] Add layerDelayedLock capability
This adds a capability to switch between layers that acts like a 'lock'
but with the difference that it requires the 'lock' key to be pressed
for 500 milliseconds before the function layer gets locked.

This is useful to avoid switching to a function layer by mistake since
pressing the lock key by accident won't actually switch layer unless
it's held for at least 500ms.