sam: Use 16k for bootloader
Was previously at 103% which could maybe squeeze into 8k, but this gives
us breathing room and the ability to add new features in the future.

Additionally, this avoids having to write to the small 8k sectors which
have different constraints than the larger 48/64k oner

The firmware is using ~20% of flash so there is little space constraint
there.