Converting GPLv3 to LGPLv3
- Needed to integrate with Nordic Softdevices
  * This will make it much easier to distribute as FCC recertification
  won't be necessary (especially for derivative works)
  * Still keeps the core libraries and KLL integration open
- The following files only have modifications from
  * Jacob Alexander (main author)
  * Rowan Decker (secondary author, has given permission for the
  relicense)

There are 3 files that need to be converted to LGPLv3
  * Macro/PartialMap/macro.c (need to acquire further permission, or
  rewrite patches)
  * Lib/entropy.c/entropy.h (need to rewrite library or find an
  alternative)
    - Fortunately the SAM version was written by Jacob
  Alexander so it shouldn't be difficult
    - No new keyboards are planned for kinetis, so no need to worry
    about handling LGPLv3 change

The Bootloader code is based off of the McHCK bootloader which is GPL.
Fortunately, the bootloader should be standalone in most cases.
And in the case of Nordic BLE devices an entirely new bootloader will be
necessary as it's not USB based.
For keyboard specific bootloader code, all has been relicensed under an
MIT license to make it easier to use in general. The overall license of
the Bootloader is still GPLv3.